Make crashkernel size configurable with reasonable defaults
Add a class parameter `crashdump::crashkernel_size` that can be set by
the user. Its default value is based on `$::memorysize_mb`: for systems
with less than 2GB of memory crashkernel=128M is enough, for systems
with 2GB to 4GB we need crashkernel=256M and for systems with more than
4GB crashkernel=512M. This might need to be extended for systems with
even more memory.

This logic might be flawed for systems with "weird" amounts of memory,
and because `$::memorysize_mb` shrinks by the amount of reserved
crashkernel memory, possibly causing a flip-flop effect when setting the
crashkernel size.

Addresses #3.