Support setting crashkernel size on Ubuntu 14.04
In Ubuntu 14.04 the kexec-tools package installs a config file
/etc/default/grub.d/kexec-tools.cfg, which is the right place to set a
proper crashkernel size. This file takes precedence over, and thus
shadows any changes made previously in, /etc/grub.d/10_linux.

Unfortunately the default crashkernel size is still way too small for
systems with more than a gigabyte of RAM so we still need to go in and
fix it.