Update memorysize thresholds for crashkernel size determination
As mentioned in a comment in params.pp the memorysize_bm shrinks by the
amount of reserved memory for the crashkernel.

The used memorysize_mb threshhold of 1800 is problematic if you use a
memorysize of 2 GB on your system (and that's a pretty common setting)
because the crashkernel_size will flap between 128M and 256M after each
reboot.

Here's the math:

If crashkernel_size is set to 256M memorysize_mb is 1745 and therefore
with next reboot the crashkernel_size is set to 128M. Then the
memorysize_mb is 1873 and on next reboot the crashkernel_size is set
again to 256M and so on.